+++ modules/core/src/main/java/org/apache/ignite/internal/ClusterMetricsSnapshot.java
@@ -20,24 +20,24 @@ package org.apache.ignite.internal;
 import java.nio.ByteBuffer;
 import java.util.Collection;
 import java.util.Map;
-import org.apache.ignite.cluster.ClusterGroup;
 import org.apache.ignite.cluster.ClusterMetrics;
 import org.apache.ignite.cluster.ClusterNode;
-import org.apache.ignite.internal.processors.cluster.NodeMetricsMessage;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.internal.S;
 import org.apache.ignite.internal.util.typedef.internal.U;
+import org.apache.ignite.plugin.extensions.communication.Message;
 import org.jetbrains.annotations.Nullable;
 
+import static java.lang.Math.max;
 import static java.lang.Math.min;
 
 /**
- * Implementation for {@link ClusterMetrics} interface.
+ * Implementation for {@link ClusterMetrics} interface which is also {@link Message}.
  * <p>
  * Note that whenever adding or removing metric parameters, care
  * must be taken to update serialize/deserialize logic as well.
  */
-public class ClusterMetricsSnapshot implements ClusterMetrics {
+public class ClusterMetricsSnapshot implements ClusterMetrics, Message {
     /** Size of serialized node metrics. */
     public static final int METRICS_SIZE =
         4/*max active jobs*/ +
@@ -96,174 +96,1000 @@ public class ClusterMetricsSnapshot implements ClusterMetrics {
         8/*current PME time*/;
 
     /** */
-    private NodeMetricsMessage m;
+    @Order(0)
+    public long lastUpdateTime = -1;
 
-    /**
-     * Creates empty snapshot.
-     */
+    /** */
+    @Order(1)
+    public int maxActiveJobs = -1;
+
+    /** */
+    @Order(2)
+    public int curActiveJobs = -1;
+
+    /** */
+    @Order(3)
+    public float avgActiveJobs = -1;
+
+    /** */
+    @Order(4)
+    public int maxWaitingJobs = -1;
+
+    /** */
+    @Order(5)
+    public int curWaitingJobs = -1;
+
+    /** */
+    @Order(6)
+    public float avgWaitingJobs = -1;
+
+    /** */
+    @Order(7)
+    public int maxRejectedJobs = -1;
+
+    /** */
+    @Order(8)
+    public int curRejectedJobs = -1;
+
+    /** */
+    @Order(9)
+    public float avgRejectedJobs = -1;
+
+    /** */
+    @Order(10)
+    public int maxCancelledJobs = -1;
+
+    /** */
+    @Order(11)
+    public int curCancelledJobs = -1;
+
+    /** */
+    @Order(12)
+    public float avgCancelledJobs = -1;
+
+    /** */
+    @Order(13)
+    public int totalRejectedJobs = -1;
+
+    /** */
+    @Order(14)
+    public int totalCancelledJobs = -1;
+
+    /** */
+    @Order(15)
+    public int totalExecutedJobs = -1;
+
+    /** */
+    @Order(16)
+    public long maxJobWaitTime = -1;
+
+    /** */
+    @Order(17)
+    public long curJobWaitTime = Long.MAX_VALUE;
+
+    /** */
+    @Order(18)
+    public double avgJobWaitTime = -1;
+
+    /** */
+    @Order(19)
+    public long maxJobExecTime = -1;
+
+    /** */
+    @Order(20)
+    public long curJobExecTime = -1;
+
+    /** */
+    @Order(21)
+    public double avgJobExecTime = -1;
+
+    /** */
+    @Order(22)
+    public int totalExecTasks = -1;
+
+    /** */
+    @Order(23)
+    public long totalIdleTime = -1;
+
+    /** */
+    @Order(24)
+    public long curIdleTime = -1;
+
+    /** */
+    @Order(25)
+    public int totalCpus = -1;
+
+    /** */
+    @Order(26)
+    public double curCpuLoad = -1;
+
+    /** */
+    @Order(27)
+    public double avgCpuLoad = -1;
+
+    /** */
+    @Order(28)
+    public double curGcCpuLoad = -1;
+
+    /** */
+    @Order(29)
+    public long heapInit = -1;
+
+    /** */
+    @Order(30)
+    public long heapUsed = -1;
+
+    /** */
+    @Order(31)
+    public long heapCommitted = -1;
+
+    /** */
+    @Order(32)
+    public long heapMax = -1;
+
+    /** */
+    @Order(33)
+    public long heapTotal = -1;
+
+    /** */
+    @Order(34)
+    public long nonHeapInit = -1;
+
+    /** */
+    @Order(35)
+    public long nonHeapUsed = -1;
+
+    /** */
+    @Order(36)
+    public long nonHeapCommitted = -1;
+
+    /** */
+    @Order(37)
+    public long nonHeapMax = -1;
+
+    /** */
+    @Order(38)
+    public long nonHeapTotal = -1;
+
+    /** */
+    @Order(39)
+    public long upTime = -1;
+
+    /** */
+    @Order(40)
+    public long startTime = -1;
+
+    /** */
+    @Order(41)
+    public long nodeStartTime = -1;
+
+    /** */
+    @Order(42)
+    public int threadCnt = -1;
+
+    /** */
+    @Order(43)
+    public int peakThreadCnt = -1;
+
+    /** */
+    @Order(44)
+    public long startedThreadCnt = -1;
+
+    /** */
+    @Order(45)
+    public int daemonThreadCnt = -1;
+
+    /** */
+    @Order(46)
+    public long lastDataVer = -1;
+
+    /** */
+    @Order(47)
+    public int sentMsgsCnt = -1;
+
+    /** */
+    @Order(48)
+    public long sentBytesCnt = -1;
+
+    /** */
+    @Order(49)
+    public int rcvdMsgsCnt = -1;
+
+    /** */
+    @Order(50)
+    public long rcvdBytesCnt = -1;
+
+    /** */
+    @Order(51)
+    public int outMesQueueSize = -1;
+
+    /** */
+    @Order(52)
+    public int totalNodes = -1;
+
+    /** */
+    @Order(53)
+    public long totalJobsExecTime = -1;
+
+    /** */
+    @Order(54)
+    public long curPmeDuration = -1;
+
+    /** Empty constructor for serialization purposes. */
     public ClusterMetricsSnapshot() {
-        m = new NodeMetricsMessage();
+        // No-op.
     }
 
     /**
-     * Creates snapshot based on the handled message.
-     */
-    public ClusterMetricsSnapshot(NodeMetricsMessage m) {
-        // As in #deserialize().
-        m.lastUpdateTime = U.currentTimeMillis();
-
-        this.m = m;
-    }
-
-    /**
-     * Create metrics for given cluster group.
+     * Create metrics for given nodes.
      *
-     * @param p Projection to get metrics for.
+     * @param nodes Nodes.
      */
-    public ClusterMetricsSnapshot(ClusterGroup p) {
-        assert p != null;
+    public ClusterMetricsSnapshot(Collection<ClusterNode> nodes) {
+        int size = nodes.size();
 
-        m = new NodeMetricsMessage(p.nodes());
+        curJobWaitTime = Long.MAX_VALUE;
+        lastUpdateTime = 0;
+        maxActiveJobs = 0;
+        curActiveJobs = 0;
+        avgActiveJobs = 0;
+        maxWaitingJobs = 0;
+        curWaitingJobs = 0;
+        avgWaitingJobs = 0;
+        maxRejectedJobs = 0;
+        curRejectedJobs = 0;
+        avgRejectedJobs = 0;
+        maxCancelledJobs = 0;
+        curCancelledJobs = 0;
+        avgCancelledJobs = 0;
+        totalRejectedJobs = 0;
+        totalCancelledJobs = 0;
+        totalExecutedJobs = 0;
+        totalJobsExecTime = 0;
+        maxJobWaitTime = 0;
+        avgJobWaitTime = 0;
+        maxJobExecTime = 0;
+        curJobExecTime = 0;
+        avgJobExecTime = 0;
+        totalExecTasks = 0;
+        totalIdleTime = 0;
+        curIdleTime = 0;
+        totalCpus = 0;
+        curCpuLoad = 0;
+        avgCpuLoad = 0;
+        curGcCpuLoad = 0;
+        heapInit = 0;
+        heapUsed = 0;
+        heapCommitted = 0;
+        heapMax = 0;
+        nonHeapInit = 0;
+        nonHeapUsed = 0;
+        nonHeapCommitted = 0;
+        nonHeapMax = 0;
+        nonHeapTotal = 0;
+        upTime = 0;
+        startTime = 0;
+        nodeStartTime = 0;
+        threadCnt = 0;
+        peakThreadCnt = 0;
+        startedThreadCnt = 0;
+        daemonThreadCnt = 0;
+        lastDataVer = 0;
+        sentMsgsCnt = 0;
+        sentBytesCnt = 0;
+        rcvdMsgsCnt = 0;
+        rcvdBytesCnt = 0;
+        outMesQueueSize = 0;
+        heapTotal = 0;
+        totalNodes = nodes.size();
+        curPmeDuration = 0;
+
+        for (ClusterNode node : nodes) {
+            ClusterMetrics m = node.metrics();
+
+            lastUpdateTime = max(lastUpdateTime, node.metrics().getLastUpdateTime());
+
+            curActiveJobs += m.getCurrentActiveJobs();
+            maxActiveJobs = max(maxActiveJobs, m.getCurrentActiveJobs());
+            avgActiveJobs += m.getCurrentActiveJobs();
+            totalExecutedJobs += m.getTotalExecutedJobs();
+            totalJobsExecTime += m.getTotalJobsExecutionTime();
+
+            totalExecTasks += m.getTotalExecutedTasks();
+
+            totalCancelledJobs += m.getTotalCancelledJobs();
+            curCancelledJobs += m.getCurrentCancelledJobs();
+            maxCancelledJobs = max(maxCancelledJobs, m.getCurrentCancelledJobs());
+            avgCancelledJobs += m.getCurrentCancelledJobs();
+
+            totalRejectedJobs += m.getTotalRejectedJobs();
+            curRejectedJobs += m.getCurrentRejectedJobs();
+            maxRejectedJobs = max(maxRejectedJobs, m.getCurrentRejectedJobs());
+            avgRejectedJobs += m.getCurrentRejectedJobs();
+
+            curWaitingJobs += m.getCurrentWaitingJobs();
+            maxWaitingJobs = max(maxWaitingJobs, m.getCurrentWaitingJobs());
+            avgWaitingJobs += m.getCurrentWaitingJobs();
+
+            maxJobExecTime = max(maxJobExecTime, m.getMaximumJobExecuteTime());
+            avgJobExecTime += m.getAverageJobExecuteTime();
+            curJobExecTime += m.getCurrentJobExecuteTime();
+
+            curJobWaitTime = min(curJobWaitTime, m.getCurrentJobWaitTime());
+            maxJobWaitTime = max(maxJobWaitTime, m.getCurrentJobWaitTime());
+            avgJobWaitTime += m.getAverageJobWaitTime();
+
+            daemonThreadCnt += m.getCurrentDaemonThreadCount();
+
+            peakThreadCnt = max(peakThreadCnt, m.getCurrentThreadCount());
+            threadCnt += m.getCurrentThreadCount();
+            startedThreadCnt += m.getTotalStartedThreadCount();
+
+            curIdleTime += m.getCurrentIdleTime();
+            totalIdleTime += m.getTotalIdleTime();
+
+            heapCommitted += m.getHeapMemoryCommitted();
+
+            heapUsed += m.getHeapMemoryUsed();
+
+            heapMax = max(heapMax, m.getHeapMemoryMaximum());
+
+            heapTotal += m.getHeapMemoryTotal();
+
+            heapInit += m.getHeapMemoryInitialized();
+
+            nonHeapCommitted += m.getNonHeapMemoryCommitted();
+
+            nonHeapUsed += m.getNonHeapMemoryUsed();
+
+            nonHeapMax = max(nonHeapMax, m.getNonHeapMemoryMaximum());
+
+            nonHeapTotal += m.getNonHeapMemoryTotal();
+
+            nonHeapInit += m.getNonHeapMemoryInitialized();
+
+            upTime = max(upTime, m.getUpTime());
+
+            lastDataVer = max(lastDataVer, m.getLastDataVersion());
+
+            sentMsgsCnt += m.getSentMessagesCount();
+            sentBytesCnt += m.getSentBytesCount();
+            rcvdMsgsCnt += m.getReceivedMessagesCount();
+            rcvdBytesCnt += m.getReceivedBytesCount();
+            outMesQueueSize += m.getOutboundMessagesQueueSize();
+
+            avgCpuLoad += m.getCurrentCpuLoad();
+
+            curPmeDuration = max(curPmeDuration, m.getCurrentPmeDuration());
+        }
+
+        curJobExecTime /= size;
+
+        avgActiveJobs /= size;
+        avgCancelledJobs /= size;
+        avgRejectedJobs /= size;
+        avgWaitingJobs /= size;
+        avgJobExecTime /= size;
+        avgJobWaitTime /= size;
+        avgCpuLoad /= size;
+
+        if (!F.isEmpty(nodes)) {
+            ClusterMetrics oldestNodeMetrics = oldest(nodes).metrics();
+
+            nodeStartTime = oldestNodeMetrics.getNodeStartTime();
+            startTime = oldestNodeMetrics.getStartTime();
+        }
+
+        Map<String, Collection<ClusterNode>> neighborhood = U.neighborhood(nodes);
+
+        curGcCpuLoad = currentGcCpuLoad(neighborhood);
+        curCpuLoad = currentCpuLoad(neighborhood);
+        totalCpus = cpuCnt(neighborhood);
+    }
+
+    /** */
+    public ClusterMetricsSnapshot(ClusterMetrics metrics) {
+        maxActiveJobs = metrics.getMaximumActiveJobs();
+        curActiveJobs = metrics.getCurrentActiveJobs();
+        avgActiveJobs = metrics.getAverageActiveJobs();
+
+        maxWaitingJobs = metrics.getMaximumWaitingJobs();
+        curWaitingJobs = metrics.getCurrentWaitingJobs();
+        avgWaitingJobs = metrics.getAverageWaitingJobs();
+
+        maxRejectedJobs = metrics.getMaximumRejectedJobs();
+        curRejectedJobs = metrics.getCurrentRejectedJobs();
+        avgRejectedJobs = metrics.getAverageRejectedJobs();
+
+        maxCancelledJobs = metrics.getMaximumCancelledJobs();
+        curCancelledJobs = metrics.getCurrentCancelledJobs();
+        avgCancelledJobs = metrics.getAverageCancelledJobs();
+
+        totalRejectedJobs = metrics.getTotalRejectedJobs();
+        totalCancelledJobs = metrics.getTotalCancelledJobs();
+        totalExecutedJobs = metrics.getTotalExecutedJobs();
+
+        maxJobWaitTime = metrics.getMaximumJobWaitTime();
+        curJobWaitTime = metrics.getCurrentJobWaitTime();
+        avgJobWaitTime = metrics.getAverageJobWaitTime();
+
+        maxJobExecTime = metrics.getMaximumJobExecuteTime();
+        curJobExecTime = metrics.getCurrentJobExecuteTime();
+        avgJobExecTime = metrics.getAverageJobExecuteTime();
+
+        totalJobsExecTime = metrics.getTotalJobsExecutionTime();
+        totalExecTasks = metrics.getTotalExecutedTasks();
+
+        curIdleTime = metrics.getCurrentIdleTime();
+        totalIdleTime = metrics.getTotalIdleTime();
+
+        totalCpus = metrics.getTotalCpus();
+        curCpuLoad = metrics.getCurrentCpuLoad();
+        avgCpuLoad = metrics.getAverageCpuLoad();
+        curGcCpuLoad = metrics.getCurrentGcCpuLoad();
+
+        heapInit = metrics.getHeapMemoryInitialized();
+        heapUsed = metrics.getHeapMemoryUsed();
+        heapCommitted = metrics.getHeapMemoryCommitted();
+        heapMax = metrics.getHeapMemoryMaximum();
+        heapTotal = metrics.getHeapMemoryTotal();
+
+        nonHeapInit = metrics.getNonHeapMemoryInitialized();
+        nonHeapUsed = metrics.getNonHeapMemoryUsed();
+        nonHeapCommitted = metrics.getNonHeapMemoryCommitted();
+        nonHeapMax = metrics.getNonHeapMemoryMaximum();
+        nonHeapTotal = metrics.getNonHeapMemoryTotal();
+
+        startTime = metrics.getStartTime();
+        nodeStartTime = metrics.getNodeStartTime();
+        upTime = metrics.getUpTime();
+
+        lastDataVer = metrics.getLastDataVersion();
+
+        curPmeDuration = metrics.getCurrentPmeDuration();
+
+        totalNodes = metrics.getTotalNodes();
+
+        threadCnt = metrics.getCurrentThreadCount();
+        peakThreadCnt = metrics.getMaximumThreadCount();
+        startedThreadCnt = metrics.getTotalStartedThreadCount();
+        daemonThreadCnt = metrics.getCurrentDaemonThreadCount();
+
+        sentMsgsCnt = metrics.getSentMessagesCount();
+        rcvdMsgsCnt = metrics.getReceivedMessagesCount();
+        outMesQueueSize = metrics.getOutboundMessagesQueueSize();
+
+        sentBytesCnt = metrics.getSentBytesCount();
+        rcvdBytesCnt = metrics.getReceivedBytesCount();
+    }
+
+    /** */
+    public static ClusterMetricsSnapshot of(ClusterMetrics metrics) {
+        return metrics instanceof ClusterMetricsSnapshot ? (ClusterMetricsSnapshot)metrics : new ClusterMetricsSnapshot(metrics);
     }
 
     /** {@inheritDoc} */
     @Override public long getHeapMemoryTotal() {
-        return m.heapMemoryTotal();
+        return heapTotal;
+    }
+
+    /**
+     * Sets total heap size.
+     *
+     * @param heapTotal Total heap.
+     */
+    public void heapMemoryTotal(long heapTotal) {
+        this.heapTotal = heapTotal;
+    }
+
+    /**
+     * Sets non-heap total heap size.
+     *
+     * @param nonHeapTotal Total heap.
+     */
+    public void nonHeapMemoryTotal(long nonHeapTotal) {
+        this.nonHeapTotal = nonHeapTotal;
     }
 
     /** {@inheritDoc} */
     @Override public long getLastUpdateTime() {
-        return m.lastUpdateTime();
+        return lastUpdateTime;
+    }
+
+    /**
+     * Sets last update time.
+     *
+     * @param lastUpdateTime Last update time.
+     */
+    public void lastUpdateTime(long lastUpdateTime) {
+        this.lastUpdateTime = lastUpdateTime;
     }
 
     /** {@inheritDoc} */
     @Override public int getMaximumActiveJobs() {
-        return m.maximumActiveJobs();
+        return maxActiveJobs;
+    }
+
+    /**
+     * Sets max active jobs.
+     *
+     * @param maxActiveJobs Max active jobs.
+     */
+    public void maximumActiveJobs(int maxActiveJobs) {
+        this.maxActiveJobs = maxActiveJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getCurrentActiveJobs() {
-        return m.currentActiveJobs();
+        return curActiveJobs;
+    }
+
+    /**
+     * Sets current active jobs.
+     *
+     * @param curActiveJobs Current active jobs.
+     */
+    public void currentActiveJobs(int curActiveJobs) {
+        this.curActiveJobs = curActiveJobs;
     }
 
     /** {@inheritDoc} */
     @Override public float getAverageActiveJobs() {
-        return m.averageActiveJobs();
+        return curActiveJobs;
+    }
+
+    /**
+     * Sets average active jobs.
+     *
+     * @param avgActiveJobs Average active jobs.
+     */
+    public void averageActiveJobs(float avgActiveJobs) {
+        this.avgActiveJobs = avgActiveJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getMaximumWaitingJobs() {
-        return m.maximumWaitingJobs();
+        return maxWaitingJobs;
+    }
+
+    /**
+     * Sets maximum waiting jobs.
+     *
+     * @param maxWaitingJobs Maximum waiting jobs.
+     */
+    public void maximumWaitingJobs(int maxWaitingJobs) {
+        this.maxWaitingJobs = maxWaitingJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getCurrentWaitingJobs() {
-        return m.currentWaitingJobs();
+        return curWaitingJobs;
+    }
+
+    /**
+     * Sets current waiting jobs.
+     *
+     * @param curWaitingJobs Current waiting jobs.
+     */
+    public void currentWaitingJobs(int curWaitingJobs) {
+        this.curWaitingJobs = curWaitingJobs;
     }
 
     /** {@inheritDoc} */
     @Override public float getAverageWaitingJobs() {
-        return m.averageWaitingJobs();
+        return avgWaitingJobs;
+    }
+
+    /**
+     * Sets average waiting jobs.
+     *
+     * @param avgWaitingJobs Average waiting jobs.
+     */
+    public void averageWaitingJobs(float avgWaitingJobs) {
+        this.avgWaitingJobs = avgWaitingJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getMaximumRejectedJobs() {
-        return m.maximumRejectedJobs();
+        return maxRejectedJobs;
+    }
+
+    /**
+     * @param maxRejectedJobs Maximum number of jobs rejected during a single collision resolution event.
+     */
+    public void maximumRejectedJobs(int maxRejectedJobs) {
+        this.maxRejectedJobs = maxRejectedJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getCurrentRejectedJobs() {
-        return m.currentRejectedJobs();
+        return curRejectedJobs;
+    }
+
+    /**
+     * @param curRejectedJobs Number of jobs rejected during most recent collision resolution.
+     */
+    public void currentRejectedJobs(int curRejectedJobs) {
+        this.curRejectedJobs = curRejectedJobs;
     }
 
     /** {@inheritDoc} */
     @Override public float getAverageRejectedJobs() {
-        return m.averageRejectedJobs();
+        return avgRejectedJobs;
+    }
+
+    /**
+     * @param avgRejectedJobs Average number of jobs this node rejects.
+     */
+    public void averageRejectedJobs(float avgRejectedJobs) {
+        this.avgRejectedJobs = avgRejectedJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getTotalRejectedJobs() {
-        return m.totalRejectedJobs();
+        return totalRejectedJobs;
+    }
+
+    /**
+     * @param totalRejectedJobs Total number of jobs this node ever rejected.
+     */
+    public void totalRejectedJobs(int totalRejectedJobs) {
+        this.totalRejectedJobs = totalRejectedJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getMaximumCancelledJobs() {
-        return m.maximumCancelledJobs();
+        return maxCancelledJobs;
+    }
+
+    /**
+     * Sets maximum cancelled jobs.
+     *
+     * @param maxCancelledJobs Maximum cancelled jobs.
+     */
+    public void maximumCancelledJobs(int maxCancelledJobs) {
+        this.maxCancelledJobs = maxCancelledJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getCurrentCancelledJobs() {
-        return m.currentCancelledJobs();
+        return curCancelledJobs;
+    }
+
+    /**
+     * Sets current cancelled jobs.
+     *
+     * @param curCancelledJobs Current cancelled jobs.
+     */
+    public void currentCancelledJobs(int curCancelledJobs) {
+        this.curCancelledJobs = curCancelledJobs;
     }
 
     /** {@inheritDoc} */
     @Override public float getAverageCancelledJobs() {
-        return m.averageCancelledJobs();
+        return avgCancelledJobs;
+    }
+
+    /**
+     * Sets average cancelled jobs.
+     *
+     * @param avgCancelledJobs Average cancelled jobs.
+     */
+    public void averageCancelledJobs(float avgCancelledJobs) {
+        this.avgCancelledJobs = avgCancelledJobs;
     }
 
     /** {@inheritDoc} */
     @Override public int getTotalExecutedJobs() {
-        return m.totalExecutedJobs();
+        return totalExecutedJobs;
+    }
+
+    /**
+     * Sets total active jobs.
+     *
+     * @param totalExecutedJobs Total active jobs.
+     */
+    public void totalExecutedJobs(int totalExecutedJobs) {
+        this.totalExecutedJobs = totalExecutedJobs;
     }
 
     /** {@inheritDoc} */
     @Override public long getTotalJobsExecutionTime() {
-        return m.totalJobsExecutionTime();
+        return totalJobsExecTime;
+    }
+
+    /**
+     * Sets total jobs execution time.
+     *
+     * @param totalJobsExecTime Total jobs execution time.
+     */
+    public void totalJobsExecutionTime(long totalJobsExecTime) {
+        this.totalJobsExecTime = totalJobsExecTime;
     }
 
     /** {@inheritDoc} */
     @Override public int getTotalCancelledJobs() {
-        return m.totalCancelledJobs();
+        return totalCancelledJobs;
+    }
+
+    /**
+     * Sets total cancelled jobs.
+     *
+     * @param totalCancelledJobs Total cancelled jobs.
+     */
+    public void totalCancelledJobs(int totalCancelledJobs) {
+        this.totalCancelledJobs = totalCancelledJobs;
     }
 
     /** {@inheritDoc} */
     @Override public long getMaximumJobWaitTime() {
-        return m.maximumJobWaitTime();
+        return maxJobWaitTime;
+    }
+
+    /**
+     * Sets max job wait time.
+     *
+     * @param maxJobWaitTime Max job wait time.
+     */
+    public void maximumJobWaitTime(long maxJobWaitTime) {
+        this.maxJobWaitTime = maxJobWaitTime;
     }
 
     /** {@inheritDoc} */
     @Override public long getCurrentJobWaitTime() {
-        return m.currentJobWaitTime();
+        return curJobWaitTime;
+    }
+
+    /**
+     * Sets current job wait time.
+     *
+     * @param curJobWaitTime Current job wait time.
+     */
+    public void currentJobWaitTime(long curJobWaitTime) {
+        this.curJobWaitTime = curJobWaitTime;
     }
 
     /** {@inheritDoc} */
     @Override public double getAverageJobWaitTime() {
-        return m.averageJobWaitTime();
+        return avgJobWaitTime;
+    }
+
+    /**
+     * Sets average job wait time.
+     *
+     * @param avgJobWaitTime Average job wait time.
+     */
+    public void averageJobWaitTime(double avgJobWaitTime) {
+        this.avgJobWaitTime = avgJobWaitTime;
     }
 
     /** {@inheritDoc} */
     @Override public long getMaximumJobExecuteTime() {
-        return m.maximumJobExecuteTime();
+        return maxJobExecTime;
+    }
+
+    /**
+     * Sets maximum job execution time.
+     *
+     * @param maxJobExecTime Maximum job execution time.
+     */
+    public void maximumJobExecuteTime(long maxJobExecTime) {
+        this.maxJobExecTime = maxJobExecTime;
     }
 
     /** {@inheritDoc} */
     @Override public long getCurrentJobExecuteTime() {
-        return m.currentJobExecuteTime();
+        return curJobExecTime;
+    }
+
+    /**
+     * Sets current job execute time.
+     *
+     * @param curJobExecTime Current job execute time.
+     */
+    public void currentJobExecuteTime(long curJobExecTime) {
+        this.curJobExecTime = curJobExecTime;
     }
 
     /** {@inheritDoc} */
     @Override public double getAverageJobExecuteTime() {
-        return m.averageJobExecuteTime();
+        return avgJobExecTime;
+    }
+
+    /**
+     * Sets average job execution time.
+     *
+     * @param avgJobExecTime Average job execution time.
+     */
+    public void averageJobExecuteTime(double avgJobExecTime) {
+        this.avgJobExecTime = avgJobExecTime;
     }
 
     /** {@inheritDoc} */
     @Override public int getTotalExecutedTasks() {
-        return m.totalExecutedTasks();
+        return totalExecTasks;
+    }
+
+    /**
+     * Sets total executed tasks count.
+     *
+     * @param totalExecTasks total executed tasks count.
+     */
+    public void totalExecutedTasks(int totalExecTasks) {
+        this.totalExecTasks = totalExecTasks;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getTotalIdleTime() {
+        return totalIdleTime;
+    }
+
+    /**
+     * Set total node idle time.
+     *
+     * @param totalIdleTime Total node idle time.
+     */
+    public void totalIdleTime(long totalIdleTime) {
+        this.totalIdleTime = totalIdleTime;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getCurrentIdleTime() {
+        return curIdleTime;
+    }
+
+    /**
+     * Sets time elapsed since execution of last job.
+     *
+     * @param curIdleTime Time elapsed since execution of last job.
+     */
+    public void currentIdleTime(long curIdleTime) {
+        this.curIdleTime = curIdleTime;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getTotalCpus() {
+        return totalCpus;
+    }
+
+    /** {@inheritDoc} */
+    @Override public double getCurrentCpuLoad() {
+        return curCpuLoad;
+    }
+
+    /** {@inheritDoc} */
+    @Override public double getAverageCpuLoad() {
+        return avgCpuLoad;
+    }
+
+    /** {@inheritDoc} */
+    @Override public double getCurrentGcCpuLoad() {
+        return curGcCpuLoad;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getHeapMemoryInitialized() {
+        return heapInit;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getHeapMemoryUsed() {
+        return heapUsed;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getHeapMemoryCommitted() {
+        return heapCommitted;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getHeapMemoryMaximum() {
+        return heapMax;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getNonHeapMemoryInitialized() {
+        return nonHeapInit;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getNonHeapMemoryUsed() {
+        return nonHeapUsed;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getNonHeapMemoryCommitted() {
+        return nonHeapCommitted;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getNonHeapMemoryMaximum() {
+        return nonHeapMax;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getNonHeapMemoryTotal() {
+        return nonHeapTotal;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getUpTime() {
+        return upTime;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getStartTime() {
+        return startTime;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getNodeStartTime() {
+        return nodeStartTime;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getCurrentThreadCount() {
+        return threadCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getMaximumThreadCount() {
+        return peakThreadCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getTotalStartedThreadCount() {
+        return startedThreadCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getCurrentDaemonThreadCount() {
+        return daemonThreadCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getLastDataVersion() {
+        return lastDataVer;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getSentMessagesCount() {
+        return sentMsgsCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getSentBytesCount() {
+        return sentBytesCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getReceivedMessagesCount() {
+        return rcvdMsgsCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getReceivedBytesCount() {
+        return rcvdBytesCnt;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getOutboundMessagesQueueSize() {
+        return outMesQueueSize;
+    }
+
+    /** {@inheritDoc} */
+    @Override public int getTotalNodes() {
+        return totalNodes;
+    }
+
+    /** {@inheritDoc} */
+    @Override public long getCurrentPmeDuration() {
+        return curPmeDuration;
     }
 
     /** {@inheritDoc} */
     @Override public long getTotalBusyTime() {
         return getUpTime() - getTotalIdleTime();
-    }
-
-    /** {@inheritDoc} */
-    @Override public long getTotalIdleTime() {
-        return m.totalIdleTime();
-    }
-
-    /** {@inheritDoc} */
-    @Override public long getCurrentIdleTime() {
-        return m.currentIdleTime();
     }
 
     /** {@inheritDoc} */
@@ -276,144 +1102,267 @@ public class ClusterMetricsSnapshot implements ClusterMetrics {
         return getTotalIdleTime() / (float)getUpTime();
     }
 
-    /** {@inheritDoc} */
-    @Override public int getTotalCpus() {
-        return m.totalCpus();
+    /**
+     * Sets available processors.
+     *
+     * @param totalCpus Available processors.
+     */
+    public void totalCpus(int totalCpus) {
+        this.totalCpus = totalCpus;
     }
 
-    /** {@inheritDoc} */
-    @Override public double getCurrentCpuLoad() {
-        return m.currentCpuLoad();
+    /**
+     * Sets current CPU load.
+     *
+     * @param curCpuLoad Current CPU load.
+     */
+    public void currentCpuLoad(double curCpuLoad) {
+        this.curCpuLoad = curCpuLoad;
     }
 
-    /** {@inheritDoc} */
-    @Override public double getAverageCpuLoad() {
-        return m.averageCpuLoad();
+    /**
+     * Sets CPU load average over the metrics history.
+     *
+     * @param avgCpuLoad CPU load average.
+     */
+    public void averageCpuLoad(double avgCpuLoad) {
+        this.avgCpuLoad = avgCpuLoad;
     }
 
-    /** {@inheritDoc} */
-    @Override public double getCurrentGcCpuLoad() {
-        return m.currentGcCpuLoad();
+    /**
+     * Sets current GC load.
+     *
+     * @param curGcCpuLoad Current GC load.
+     */
+    public void currentGcCpuLoad(double curGcCpuLoad) {
+        this.curGcCpuLoad = curGcCpuLoad;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getHeapMemoryInitialized() {
-        return m.heapMemoryInitialized();
+    /**
+     * Sets heap initial memory.
+     *
+     * @param heapInit Heap initial memory.
+     */
+    public void heapMemoryInitialized(long heapInit) {
+        this.heapInit = heapInit;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getHeapMemoryUsed() {
-        return m.heapMemoryUsed();
+    /**
+     * Sets used heap memory.
+     *
+     * @param heapUsed Used heap memory.
+     */
+    public void heapMemoryUsed(long heapUsed) {
+        this.heapUsed = heapUsed;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getHeapMemoryCommitted() {
-        return m.heapMemoryCommitted();
+    /**
+     * Sets committed heap memory.
+     *
+     * @param heapCommitted Committed heap memory.
+     */
+    public void heapMemoryCommitted(long heapCommitted) {
+        this.heapCommitted = heapCommitted;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getHeapMemoryMaximum() {
-        return m.heapMemoryMaximum();
+    /**
+     * Sets maximum possible heap memory.
+     *
+     * @param heapMax Maximum possible heap memory.
+     */
+    public void heapMemoryMaximum(long heapMax) {
+        this.heapMax = heapMax;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getNonHeapMemoryInitialized() {
-        return m.nonHeapMemoryInitialized();
+    /**
+     * Sets initial non-heap memory.
+     *
+     * @param nonHeapInit Initial non-heap memory.
+     */
+    public void nonHeapMemoryInitialized(long nonHeapInit) {
+        this.nonHeapInit = nonHeapInit;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getNonHeapMemoryUsed() {
-        return m.nonHeapMemoryUsed();
+    /**
+     * Sets used non-heap memory.
+     *
+     * @param nonHeapUsed Used non-heap memory.
+     */
+    public void nonHeapMemoryUsed(long nonHeapUsed) {
+        this.nonHeapUsed = nonHeapUsed;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getNonHeapMemoryCommitted() {
-        return m.nonHeapMemoryCommitted();
+    /**
+     * Sets committed non-heap memory.
+     *
+     * @param nonHeapCommitted Committed non-heap memory.
+     */
+    public void nonHeapMemoryCommitted(long nonHeapCommitted) {
+        this.nonHeapCommitted = nonHeapCommitted;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getNonHeapMemoryMaximum() {
-        return m.nonHeapMemoryMaximum();
+    /**
+     * Sets maximum possible non-heap memory.
+     *
+     * @param nonHeapMax Maximum possible non-heap memory.
+     */
+    public void nonHeapMemoryMaximum(long nonHeapMax) {
+        this.nonHeapMax = nonHeapMax;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getNonHeapMemoryTotal() {
-        return m.nonHeapMemoryTotal();
+    /**
+     * Sets VM up time.
+     *
+     * @param upTime VM up time.
+     */
+    public void upTime(long upTime) {
+        this.upTime = upTime;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getUpTime() {
-        return m.upTime();
+    /**
+     * Sets VM start time.
+     *
+     * @param startTime VM start time.
+     */
+    public void startTime(long startTime) {
+        this.startTime = startTime;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getStartTime() {
-        return m.startTime();
+    /**
+     * Sets node start time.
+     *
+     * @param nodeStartTime node start time.
+     */
+    public void nodeStartTime(long nodeStartTime) {
+        this.nodeStartTime = nodeStartTime;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getNodeStartTime() {
-        return m.nodeStartTime();
+    /**
+     * Sets thread count.
+     *
+     * @param threadCnt Thread count.
+     */
+    public void currentThreadCount(int threadCnt) {
+        this.threadCnt = threadCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getCurrentThreadCount() {
-        return m.currentThreadCount();
+    /**
+     * Sets peak thread count.
+     *
+     * @param peakThreadCnt Peak thread count.
+     */
+    public void maximumThreadCount(int peakThreadCnt) {
+        this.peakThreadCnt = peakThreadCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getMaximumThreadCount() {
-        return m.maximumThreadCount();
+    /**
+     * Sets started thread count.
+     *
+     * @param startedThreadCnt Started thread count.
+     */
+    public void totalStartedThreadCount(long startedThreadCnt) {
+        this.startedThreadCnt = startedThreadCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getTotalStartedThreadCount() {
-        return m.totalStartedThreadCount();
+    /**
+     * Sets daemon thread count.
+     *
+     * @param daemonThreadCnt Daemon thread count.
+     */
+    public void currentDaemonThreadCount(int daemonThreadCnt) {
+        this.daemonThreadCnt = daemonThreadCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getCurrentDaemonThreadCount() {
-        return m.currentDaemonThreadCount();
+    /**
+     * Sets last data version.
+     *
+     * @param lastDataVer Last data version.
+     */
+    public void lastDataVersion(long lastDataVer) {
+        this.lastDataVer = lastDataVer;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getLastDataVersion() {
-        return m.lastDataVersion();
+    /**
+     * Sets sent messages count.
+     *
+     * @param sentMsgsCnt Sent messages count.
+     */
+    public void sentMessagesCount(int sentMsgsCnt) {
+        this.sentMsgsCnt = sentMsgsCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getSentMessagesCount() {
-        return m.sentMessagesCount();
+    /**
+     * Sets sent bytes count.
+     *
+     * @param sentBytesCnt Sent bytes count.
+     */
+    public void sentBytesCount(long sentBytesCnt) {
+        this.sentBytesCnt = sentBytesCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getSentBytesCount() {
-        return m.sentBytesCount();
+    /**
+     * Sets received messages count.
+     *
+     * @param rcvdMsgsCnt Received messages count.
+     */
+    public void receivedMessagesCount(int rcvdMsgsCnt) {
+        this.rcvdMsgsCnt = rcvdMsgsCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getReceivedMessagesCount() {
-        return m.receivedMessagesCount();
+    /**
+     * Sets received bytes count.
+     *
+     * @param rcvdBytesCnt Received bytes count.
+     */
+    public void receivedBytesCount(long rcvdBytesCnt) {
+        this.rcvdBytesCnt = rcvdBytesCnt;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getReceivedBytesCount() {
-        return m.receivedBytesCount();
+    /**
+     * Sets outbound messages queue size.
+     *
+     * @param outMesQueueSize Outbound messages queue size.
+     */
+    public void outboundMessagesQueueSize(int outMesQueueSize) {
+        this.outMesQueueSize = outMesQueueSize;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getOutboundMessagesQueueSize() {
-        return m.outboundMessagesQueueSize();
+    /**
+     * Sets total number of nodes.
+     *
+     * @param totalNodes Total number of nodes.
+     */
+    public void totalNodes(int totalNodes) {
+        this.totalNodes = totalNodes;
     }
 
-    /** {@inheritDoc} */
-    @Override public int getTotalNodes() {
-        return m.totalNodes();
+    /**
+     * Sets execution duration for current partition map exchange.
+     *
+     * @param curPmeDuration Execution duration for current partition map exchange.
+     */
+    public void currentPmeDuration(long curPmeDuration) {
+        this.curPmeDuration = curPmeDuration;
     }
 
-    /** {@inheritDoc} */
-    @Override public long getCurrentPmeDuration() {
-        return m.currentPmeDuration();
+    /**
+     * Gets the oldest node in given collection.
+     *
+     * @param nodes Nodes.
+     * @return Oldest node or {@code null} if collection is empty.
+     */
+    @Nullable private static ClusterNode oldest(Collection<ClusterNode> nodes) {
+        long min = Long.MAX_VALUE;
+
+        ClusterNode oldest = null;
+
+        for (ClusterNode n : nodes)
+            if (n.order() < min) {
+                min = n.order();
+                oldest = n;
+            }
+
+        return oldest;
     }
 
     /**
@@ -438,56 +1387,36 @@ public class ClusterMetricsSnapshot implements ClusterMetrics {
      * @param neighborhood Cluster neighborhood.
      * @return CPU load.
      */
-    private static int cpus(Map<String, Collection<ClusterNode>> neighborhood) {
-        int cpus = 0;
+    private static double currentCpuLoad(Map<String, Collection<ClusterNode>> neighborhood) {
+        double curCpuLoad = 0.0;
 
         for (Collection<ClusterNode> nodes : neighborhood.values()) {
             ClusterNode first = F.first(nodes);
 
             // Projection can be empty if all nodes in it failed.
             if (first != null)
-                cpus += first.metrics().getCurrentCpuLoad();
+                curCpuLoad += first.metrics().getCurrentCpuLoad();
         }
 
-        return cpus;
+        return curCpuLoad;
     }
 
     /**
      * @param neighborhood Cluster neighborhood.
      * @return GC CPU load.
      */
-    private static int gcCpus(Map<String, Collection<ClusterNode>> neighborhood) {
-        int cpus = 0;
+    private static double currentGcCpuLoad(Map<String, Collection<ClusterNode>> neighborhood) {
+        double curGcCpuLoad = 0;
 
         for (Collection<ClusterNode> nodes : neighborhood.values()) {
             ClusterNode first = F.first(nodes);
 
             // Projection can be empty if all nodes in it failed.
             if (first != null)
-                cpus += first.metrics().getCurrentGcCpuLoad();
+                curGcCpuLoad += first.metrics().getCurrentGcCpuLoad();
         }
 
-        return cpus;
-    }
-
-    /**
-     * Gets the oldest node in given collection.
-     *
-     * @param nodes Nodes.
-     * @return Oldest node or {@code null} if collection is empty.
-     */
-    @Nullable private static ClusterNode oldest(Collection<ClusterNode> nodes) {
-        long min = Long.MAX_VALUE;
-
-        ClusterNode oldest = null;
-
-        for (ClusterNode n : nodes)
-            if (n.order() < min) {
-                min = n.order();
-                oldest = n;
-            }
-
-        return oldest;
+        return curGcCpuLoad;
     }
 
     /**
@@ -583,8 +1512,8 @@ public class ClusterMetricsSnapshot implements ClusterMetrics {
      * @param off Offset into byte array.
      * @return Deserialized node metrics.
      */
-    public static ClusterMetrics deserialize(byte[] data, int off) {
-        NodeMetricsMessage msg = new NodeMetricsMessage();
+    public static ClusterMetricsSnapshot deserialize(byte[] data, int off) {
+        ClusterMetricsSnapshot msg = new ClusterMetricsSnapshot();
 
         int bufSize = min(METRICS_SIZE, data.length - off);
 
@@ -656,7 +1585,7 @@ public class ClusterMetricsSnapshot implements ClusterMetrics {
         else
             msg.currentPmeDuration(0);
 
-        return new ClusterMetricsSnapshot(msg);
+        return msg;
     }
 
     /** {@inheritDoc} */
