+++ modules/core/src/main/java/org/apache/ignite/internal/CoreMessagesProvider.java
@@ -200,7 +200,6 @@ import org.apache.ignite.internal.processors.cluster.ClusterMetricsUpdateMessage
 import org.apache.ignite.internal.processors.cluster.ClusterUpdateNotifierDataBagItem;
 import org.apache.ignite.internal.processors.cluster.DiscoveryDataClusterState;
 import org.apache.ignite.internal.processors.cluster.NodeFullMetricsMessage;
-import org.apache.ignite.internal.processors.cluster.NodeMetricsMessage;
 import org.apache.ignite.internal.processors.continuous.ContinuousRoutineStartResultMessage;
 import org.apache.ignite.internal.processors.continuous.GridContinuousMessage;
 import org.apache.ignite.internal.processors.continuous.StartRequestData;
@@ -666,7 +665,7 @@ public class CoreMessagesProvider extends AbstractMessageFactoryProvider {
         // [11900 - 12000]: Metrics, monitoring messages.
         msgIdx = 11900;
         register(CacheMetricsMessage.class);
-        register(NodeMetricsMessage.class);
+        register(ClusterMetricsSnapshot.class);
         register(NodeFullMetricsMessage.class);
         register(ClusterMetricsUpdateMessage.class);
         register(TcpDiscoveryClientNodesMetricsMessage.class);
